Refactor KoliBri components to align with Skeleton architecture
- Updated AGENTS.md and ARC42.md to clarify architectural guidelines and props-first refactoring workflow.
- Introduced REFACTORING_PROMPT.md for structured component migration instructions.
- Migrated props to dedicated files under src/internal/props/ for better organization and type safety.
- Implemented new snapshot and interaction tests for KolClickButton and KolSkeleton components.
- Removed outdated snapshot and test files to maintain a clean codebase.
- Enhanced documentation for controller state management and event handler conventions.
